# More optimizations, remove history preservation