--- a/Sources/_StringProcessing/Engine/MECapture.swift
+++ b/Sources/_StringProcessing/Engine/MECapture.swift
@@ -32,59 +32,48 @@
 
 extension Processor {
   struct _StoredCapture {
-    // Set whenever we push the very first capture, allows us
-    // to theoretically re-compute anything we want to later.
-    fileprivate var startState: SavePoint? = nil
+    var range: Range<Position>? = nil
 
-    // Save the entire history as we go, so that backtracking
-    // can just lop-off aborted runs.
-    //
-    // Backtracking entries can specify a per-capture stack
-    // index so that we can abort anything that came after.
-    //
-    // By remembering the entire history, we waste space, but
-    // we get flexibility for now.
-    //
-    fileprivate var history: Array<(range: Range<Position>, value: Any?)> = []
+    var value: Any? = nil
 
     // An in-progress capture start
     fileprivate var currentCaptureBegin: Position? = nil
 
     fileprivate func _invariantCheck() {
-      if startState == nil {
-        assert(history.isEmpty)
-        assert(currentCaptureBegin == nil)
-      } else if currentCaptureBegin == nil {
-        assert(!history.isEmpty)
+      if range == nil {
+        assert(value == nil)
       }
     }
 
     // MARK: - IPI
 
-    var isEmpty: Bool { history.isEmpty }
-
-    var latest: (range: Range<Position>, value: Any?)? { history.last }
+    var deconstructed: (range: Range<Position>, value: Any?)? {
+      guard let r = range else { return nil }
+      return (r, value)
+    }
 
     /// Start a new capture. If the previously started one was un-ended,
-    /// will clear it and restart. If this is the first start, will save `initial`.
+    /// will clear it and restart.
     mutating func startCapture(
-      _ idx: Position, initial: SavePoint
+      _ idx: Position
     ) {
       _invariantCheck()
       defer { _invariantCheck() }
 
-      if self.startState == nil {
-        self.startState = initial
-      }
       currentCaptureBegin = idx
     }
 
     mutating func endCapture(_ idx: Position) {
       _invariantCheck()
-      assert(currentCaptureBegin != nil)
       defer { _invariantCheck() }
 
-      history.append((currentCaptureBegin! ..< idx, value: nil))
+      guard let low = currentCaptureBegin else {
+        fatalError("Invariant violated: ending unstarted capture")
+      }
+
+      range = low..<idx
+      value = nil // TODO: cleaner IPI around this...
+      currentCaptureBegin = nil
     }
 
     mutating func registerValue(
@@ -93,28 +82,15 @@ extension Processor {
     ) {
       _invariantCheck()
       defer { _invariantCheck() }
-      if let sp = overwriteInitial {
-        self.startState = sp
-      }
-      history[history.endIndex - 1].value = value
-    }
 
-    mutating func fail(truncatingAt stackIdx: Int) {
-      _invariantCheck()
-      assert(stackIdx <= history.endIndex)
-      defer { _invariantCheck() }
-
-      history.removeSubrange(stackIdx...)
-      if history.isEmpty {
-        startState = nil
-      }
+      self.value = value
     }
   }
 }
 
 extension Processor._StoredCapture: CustomStringConvertible {
   var description: String {
-    return String(describing: history)
+    return String(describing: self)
   }
 }
 
@@ -124,10 +100,10 @@ struct MECaptureList {
 
   func latestUntyped(from input: String) -> Array<Substring?> {
     values.map {
-      guard let last = $0.latest else {
+      guard let range = $0.range else {
         return nil
       }
-      return input[last.0]
+      return input[range]
     }
   }
 }