--- a/Sources/_StringProcessing/Regex/Match.swift
+++ b/Sources/_StringProcessing/Regex/Match.swift
@@ -140,14 +140,19 @@ extension Regex {
     // FIXME: Something more efficient, likely an engine interface, and we
     // should scrap the RegexConsumer crap and call this
 
+    let executor = Executor(program: regex.program.loweredProgram)
+    let graphemeSemantic = regex.initialOptions.semanticLevel == .graphemeCluster
+
     var low = inputRange.lowerBound
     let high = inputRange.upperBound
     while true {
-      if let m = try _match(input, in: low..<high, mode: .partialFromFront) {
+      if let m: Regex<Output>.Match = try executor.match(
+        input, in: low..<high, .partialFromFront
+      ) {
         return m
       }
       if low >= high { return nil }
-      if regex.initialOptions.semanticLevel == .graphemeCluster {
+      if graphemeSemantic {
         input.formIndex(after: &low)
       } else {
         input.unicodeScalars.formIndex(after: &low)