Implement some more of the identified optimizations:

```
- ReluctantQuant 176ms
- ReluctantQuantWithTerminal 231ms
- EagarQuantWithTerminal 69.1ms
- BasicBacktrack 20.5ms
- BasicBacktrackFirstMatch 19.9ms
- cssRegex 93.6ms
- FirstMatch 140ms
- AllMatches 37.1ms

After

- ReluctantQuant 117ms
- ReluctantQuantWithTerminal 75ms
- EagarQuantWithTerminal 68.3ms
- BasicBacktrack 10ms
- BasicBacktrackFirstMatch 10.1ms
- cssRegex 41.4ms
- FirstMatch 76.8ms
- AllMatches 38ms

```